test(app-opine): consolidate crawler tests
